+++ x/evidence/types/codec.go
@@ -1,7 +1,6 @@
 package types
 
 import (
-	corelegacy "cosmossdk.io/core/legacy"
 	"cosmossdk.io/core/registry"
 	coretransaction "cosmossdk.io/core/transaction"
 	"cosmossdk.io/x/evidence/exported"
@@ -12,10 +11,10 @@ import (
 
 // RegisterLegacyAminoCodec registers all the necessary types and interfaces for the
 // evidence module.
-func RegisterLegacyAminoCodec(cdc corelegacy.Amino) {
-	cdc.RegisterInterface((*exported.Evidence)(nil), nil)
-	legacy.RegisterAminoMsg(cdc, &MsgSubmitEvidence{}, "cosmos-sdk/MsgSubmitEvidence")
-	cdc.RegisterConcrete(&Equivocation{}, "cosmos-sdk/Equivocation")
+func RegisterLegacyAminoCodec(registrar registry.AminoRegistrar) {
+	registrar.RegisterInterface((*exported.Evidence)(nil), nil)
+	legacy.RegisterAminoMsg(registrar, &MsgSubmitEvidence{}, "cosmos-sdk/MsgSubmitEvidence")
+	registrar.RegisterConcrete(&Equivocation{}, "cosmos-sdk/Equivocation")
 }
 
 // RegisterInterfaces registers the interfaces types with the interface registry.