+++ x/upgrade/depinject.go
@@ -6,11 +6,10 @@ import (
 
 	modulev1 "cosmossdk.io/api/cosmos/upgrade/module/v1"
 	"cosmossdk.io/core/address"
-	"cosmossdk.io/core/app"
 	"cosmossdk.io/core/appmodule"
+	coreserver "cosmossdk.io/core/server"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/depinject/appconfig"
-	authtypes "cosmossdk.io/x/auth/types"
 	"cosmossdk.io/x/upgrade/keeper"
 	"cosmossdk.io/x/upgrade/types"
 
@@ -19,6 +18,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/server"
 	servertypes "github.com/cosmos/cosmos-sdk/server/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 )
 
 var _ depinject.OnePerModuleType = AppModule{}
@@ -40,7 +40,8 @@ type ModuleInputs struct {
 	Environment        appmodule.Environment
 	Cdc                codec.Codec
 	AddressCodec       address.Codec
-	AppVersionModifier app.VersionModifier
+	AppVersionModifier coreserver.VersionModifier
+	ConsensusKeeper    types.ConsensusKeeper
 
 	AppOpts servertypes.AppOptions `optional:"true"` // server v0
 	Viper   *viper.Viper           `optional:"true"` // server v2
@@ -85,7 +86,7 @@ func ProvideModule(in ModuleInputs) ModuleOutputs {
 	}
 
 	// set the governance module account as the authority for conducting upgrades
-	k := keeper.NewKeeper(in.Environment, skipUpgradeHeights, in.Cdc, homePath, in.AppVersionModifier, authorityStr)
+	k := keeper.NewKeeper(in.Environment, skipUpgradeHeights, in.Cdc, homePath, in.AppVersionModifier, authorityStr, in.ConsensusKeeper)
 	m := NewAppModule(k)
 
 	return ModuleOutputs{UpgradeKeeper: k, Module: m}